Migrate entity tests
